Fix containerd overwrite problem

+++ parts/linux/cloud-init/artifacts/cse_config.sh
@@ -345,22 +345,27 @@ LimitNOFILE=1048576
 EOF
 
   mkdir -p /etc/containerd
-  # Remove in case this is an existing symlink
-  rm -f /etc/containerd/config.toml
-  if [ "${GPU_NODE}" = "true" ]; then
-    # Check VM tag directly to determine if GPU drivers should be skipped
-    export -f should_skip_nvidia_drivers
-    should_skip=$(should_skip_nvidia_drivers)
-    if [ "$?" -eq 0 ] && [ "${should_skip}" = "true" ]; then
-      echo "Generating non-GPU containerd config for GPU node due to VM tags"
-      echo "${CONTAINERD_CONFIG_NO_GPU_CONTENT}" | base64 -d > /etc/containerd/config.toml || exit $ERR_FILE_WATCH_TIMEOUT
+
+  if grep -q 'BinaryName = "/usr/bin/nvidia-container-runtime"' /etc/containerd/config.toml 2>/dev/null; then
+    echo "NVIDIA containerd config already exists at /etc/containerd/config.toml, skipping generation"
+  else
+    # Remove in case this is an existing symlink or non-NVIDIA config
+    rm -f /etc/containerd/config.toml
+    if [ "${GPU_NODE}" = "true" ]; then
+      # Check VM tag directly to determine if GPU drivers should be skipped
+      export -f should_skip_nvidia_drivers
+      should_skip=$(should_skip_nvidia_drivers)
+      if [ "$?" -eq 0 ] && [ "${should_skip}" = "true" ]; then
+        echo "Generating non-GPU containerd config for GPU node due to VM tags"
+        echo "${CONTAINERD_CONFIG_NO_GPU_CONTENT}" | base64 -d > /etc/containerd/config.toml || exit $ERR_FILE_WATCH_TIMEOUT
+      else
+        echo "Generating GPU containerd config..."
+        echo "${CONTAINERD_CONFIG_CONTENT}" | base64 -d > /etc/containerd/config.toml || exit $ERR_FILE_WATCH_TIMEOUT
+      fi
     else
-      echo "Generating GPU containerd config..."
+      echo "Generating containerd config..."
       echo "${CONTAINERD_CONFIG_CONTENT}" | base64 -d > /etc/containerd/config.toml || exit $ERR_FILE_WATCH_TIMEOUT
     fi
-  else
-    echo "Generating containerd config..."
-    echo "${CONTAINERD_CONFIG_CONTENT}" | base64 -d > /etc/containerd/config.toml || exit $ERR_FILE_WATCH_TIMEOUT
   fi
 
   export -f should_e2e_mock_azure_china_cloud
